Replace long with int in RemoteFile.getLastModified()
It seems that only ints are passed in and out...

+++ src/se/vidstige/jadb/RemoteFile.java
@@ -10,7 +10,7 @@ public class RemoteFile {
 
     public String getName() { throw new UnsupportedOperationException(); }
     public int getSize() { throw new UnsupportedOperationException(); }
-    public long getLastModified() { throw new UnsupportedOperationException(); }
+    public int getLastModified() { throw new UnsupportedOperationException(); }
     public boolean isDirectory() { throw new UnsupportedOperationException(); }
 
     public String getPath() { return path;}
